backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
plexapi
pydantic
python-dotenv
alembic
requests
apscheduler
